conscious-cart-coach/requirements.txt
# Data Processing
pandas>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# LLM Integration
anthropic>=0.18.0
# openai>=1.0.0  # Alternative to anthropic

# Observability & Evaluation
opik>=0.1.0

# Web UI
streamlit>=1.30.0

# Configuration
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-cov>=4.0.0
